Yo dawg, heard you had broken dependencies so I put broken dependencies in your broken dependencies
We were attempting to use Vault as a dependency, but it wasn't added. After playing around for a bit, I finally found a current, existing source for where the API is currently hosted.

If all is well in the world, this should build properly. I TESTED IT LOCALLY. PLS WORK.